refactor: file and folder restructure
